test(signing): submit only psbt to finalize signer
